requirements.txt
psutil
pefile
pyopenssl
certifi
tqdm
ipaddress
requests